Data Packing attempt

+++ labs/memory_bound/data_packing/init.cpp
@@ -7,7 +7,7 @@ S create_entry(int first_value, int second_value) {
 
   entry.i = first_value;
   entry.s = second_value;
-  entry.l = static_cast<uint16_t>(first_value * second_value);
+  entry.l = static_cast<uint32_t>(first_value * second_value);
   entry.d = static_cast<float>(first_value) / maxRandom;
   entry.b = first_value < second_value;
